refactor: optimize code cleanliness and image path compliance
