+++ src/modules/product-details/components/ProductDetails/ProductDetails.module.scss
@@ -6,38 +6,6 @@
   color: $text-main;
 }
 
-.backButton {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-  margin-top: 24px;
-  border: none;
-  cursor: pointer;
-  background-color: transparent;
-  color: $text-main;
-  padding: 0;
-  transition: opacity 0.3s ease;
-
-  @include small-text;
-
-  &:hover {
-    opacity: 0.7;
-  }
-
-  @media (min-width: 768px) {
-    margin-top: 40px;
-  }
-}
-
-.back {
-  margin: 0;
-  border: none;
-  display: block;
-  transition: filter 0.3s ease;
-  filter: $filter;
-}
-
 .title {
   flex: auto;
   color: $text-main;
